🚚 Move char arbitraries into /arbitrary
