+++ src/arbitrary/hexa.ts
@@ -1,0 +1,18 @@
+import { Arbitrary } from '../check/arbitrary/definition/Arbitrary';
+import { buildCharacterArbitrary } from './_internals/builders/CharacterArbitraryBuilder';
+
+/** @internal */
+function hexaMapper(v: number): number {
+  return v < 10
+    ? v + 48 // 0-9
+    : v + 97 - 10; // a-f
+}
+
+/**
+ * For single hexadecimal characters - 0-9 or a-f
+ * @remarks Since 0.0.1
+ * @public
+ */
+export function hexa(): Arbitrary<string> {
+  return buildCharacterArbitrary(0, 15, hexaMapper);
+}
