+++ src/fast-check-default.ts
@@ -26,7 +26,13 @@ import { bigUint, BigUintConstraints } from './arbitrary/bigUint';
 import { bigUintN } from './arbitrary/bigUintN';
 import { boolean } from './check/arbitrary/BooleanArbitrary';
 import { falsy, FalsyContraints, FalsyValue } from './arbitrary/falsy';
-import { ascii, base64, char, char16bits, fullUnicode, hexa, unicode } from './check/arbitrary/CharacterArbitrary';
+import { ascii } from './arbitrary/ascii';
+import { base64 } from './arbitrary/base64';
+import { char } from './arbitrary/char';
+import { char16bits } from './arbitrary/char16bits';
+import { fullUnicode } from './arbitrary/fullUnicode';
+import { hexa } from './arbitrary/hexa';
+import { unicode } from './arbitrary/unicode';
 import { clonedConstant } from './arbitrary/clonedConstant';
 import { constant } from './arbitrary/constant';
 import { constantFrom } from './arbitrary/constantFrom';
